add fuentes

+++ 14 ejercicios/pyton/ejer11.py
@@ -1,5 +1,5 @@
-r1=int(input("Escriba una resistencia"))
-r2=int(input("Escriba otra resistencia"))
+r1=float(input("Escriba una resistencia"))
+r2=float(input("Escriba otra resistencia"))
 
 req = str((r1*r2)/(r1+r2))
